Remove recent mlx4 patches
- Delete
  patches.drivers/0001-mlx4_en-Adding-40gb-speed-report-for-ethtool.patch.
- Delete
  patches.drivers/0002-mlx4_core-Extend-capability-flags-to-64-bits.patch.
- Delete
  patches.drivers/0003-mlx4-capability-for-link-sensing.patch.
- Delete
  patches.drivers/0004-mlx4_core-Stash-PCI-ID-driver_data-in-mlx4_priv-str.patch.
- Delete
  patches.drivers/0005-mlx4_core-Clean-up-enabling-of-SENSE_PORT-for-older.patch.
- Delete
  patches.drivers/0006-mlx4_en-fix-skb-truesize-underestimation.patch.
- Delete
  patches.drivers/0007-mlx4_en-nullify-cached-multicast-address-list-after.patch.
- Delete
  patches.drivers/0008-mlx4-allocate-just-enough-pages-instead-of-always-4.patch.
- Delete
  patches.drivers/0009-net-mlx4_core-Remove-port-type-restrictions.patch.

suse-commit: ef4ad8dfe5daedbe21a4d2a9ccd83fb1617d2fe3

+++ drivers/net/mlx4/fw.h
@@ -83,7 +83,7 @@ struct mlx4_dev_cap {
 	int vep_uc_steering;
 	int vep_mc_steering;
 	int wol;
-	u64 flags;
+	u32 flags;
 	int reserved_uars;
 	int uar_size;
 	int min_page_sz;
